+++ src/lib/Scenes/City/City.tsx
@@ -1,4 +1,5 @@
 import { Box, color, Flex, Theme } from "@artsy/palette"
+import { Schema, screenTrack } from "lib/utils/track"
 import React, { Component } from "react"
 import { NativeModules, ScrollView } from "react-native"
 import { RelayProp } from "react-relay"
@@ -13,6 +14,8 @@ import { EventList } from "./Components/EventList"
 interface Props {
   verticalMargin?: number
   isDrawerOpen?: boolean
+  citySlug: string
+  tracking: any
 }
 
 interface State {
@@ -24,6 +27,29 @@ interface State {
   sponsoredContent: { introText: string; artGuideUrl: string }
 }
 
+const screenSchemaForCurrentTabState = currentSelectedTab => {
+  switch (currentSelectedTab) {
+    case "all":
+      return Schema.PageNames.CityGuideAllGuide
+    case "saved":
+      return Schema.PageNames.CityGuideSavedGuide
+    case "fairs":
+      return Schema.PageNames.CityGuideFairsGuide
+    case "galleries":
+      return Schema.PageNames.CityGuideGalleriesGuide
+    case "museums":
+      return Schema.PageNames.CityGuideMuseumsGuide
+    default:
+      return null
+  }
+}
+
+@screenTrack<Props>(props => ({
+  context_screen: screenSchemaForCurrentTabState("all"),
+  context_screen_owner_type: Schema.OwnerEntityTypes.CityGuide,
+  context_screen_owner_slug: props.citySlug,
+  context_screen_owner_id: props.citySlug,
+}))
 export class CityView extends Component<Props, State> {
   state = {
     buckets: null,
@@ -77,6 +103,12 @@ export class CityView extends Component<Props, State> {
     EventEmitter.unsubscribe("map:change", this.handleEvent)
   }
 
+  componentWillReceiveProps(nextProps) {
+    if (this.props.isDrawerOpen !== nextProps.isDrawerOpen) {
+      this.fireScreenViewAnalytics()
+    }
+  }
+
   componentDidUpdate() {
     if (!this.props.isDrawerOpen && this.scrollView) {
       this.scrollView.scrollTo({ x: 0, y: 0, animated: true })
@@ -86,6 +118,15 @@ export class CityView extends Component<Props, State> {
       // We have the Relay response; post a notification so that the ARMapContainerViewController can finalize the native UI.
       NativeModules.ARNotificationsManager.postNotificationName("ARLocalDiscoveryQueryResponseReceived", {})
     }
+  }
+
+  fireScreenViewAnalytics = () => {
+    this.props.tracking.trackEvent({
+      context_screen: screenSchemaForCurrentTabState(this.state.filter.id),
+      context_screen_owner_type: Schema.OwnerEntityTypes.CityGuide,
+      context_screen_owner_slug: this.state.citySlug,
+      context_screen_owner_id: this.state.citySlug,
+    })
   }
 
   render() {