+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 - Decrease likelihood of getting a blank view while scrolling fast in city tabs - alloy
 - Hook up save functionality for show card - luc
 - Adds loading screen for Map view - luc
+- Adds analytics to map and city view - kieran
 
 ### 1.8.14